14-post-message-back/pyproject.toml
[project]
name = "my-bot"
version = "0.1.0"
description = "Step 12: Cron + Heartbeat - Scheduled Tasks"
requires-python = ">=3.11"

dependencies = [
    "litellm>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "crawl4ai>=0.3.0",
    "watchdog>=3.0.0",
    "python-telegram-bot>=20.0",
    "discord.py>=2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "croniter>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mybot"]

[project.scripts]
my-bot = "mybot.cli.main:app"
